refactor: change default login function to blueking and delete so login function issue #3759
