+++ src/common/definitions.go
@@ -960,7 +960,8 @@ const (
 )
 
 const (
-	BKDefaultLoginUserPluginVersion = "self"
+	// BKDefaultLoginUserPluginVersion 默认的登录方式
+	BKDefaultLoginUserPluginVersion = "blueking"
 	HTTPCookieBKToken               = "bk_token"
 
 	WEBSessionUinKey           = "username"
