+++ src/web_server/middleware/login.go
@@ -13,7 +13,6 @@
 package middleware
 
 import (
-	"plugin"
 	"strings"
 
 	"configcenter/src/apimachinery/discovery"
@@ -33,7 +32,6 @@ import (
 
 var Engine *backbone.Engine
 var CacheCli *redis.Client
-var LoginPlg *plugin.Plugin
 
 // ValidLogin valid the user login status
 func ValidLogin(config options.Config, disc discovery.DiscoveryInterface) gin.HandlerFunc {
@@ -88,7 +86,7 @@ func ValidLogin(config options.Config, disc discovery.DiscoveryInterface) gin.Ha
 				c.Abort()
 				return
 			} else {
-				user := user.NewUser(config, Engine, CacheCli, LoginPlg)
+				user := user.NewUser(config, Engine, CacheCli)
 				url := user.GetLoginUrl(c)
 				c.Redirect(302, url)
 				c.Abort()
@@ -102,29 +100,7 @@ func ValidLogin(config options.Config, disc discovery.DiscoveryInterface) gin.Ha
 // IsAuthed check user is authed
 func isAuthed(c *gin.Context, config options.Config) bool {
 	rid := util.GetHTTPCCRequestID(c.Request.Header)
-	if "1" == config.Session.Skip {
-		session := sessions.Default(c)
-
-		cookieOwnerID, err := c.Cookie(common.BKHTTPOwnerID)
-		if "" == cookieOwnerID || nil != err {
-			c.SetCookie(common.BKHTTPOwnerID, common.BKDefaultOwnerID, 0, "/", "", false, false)
-			session.Set(common.WEBSessionOwnerUinKey, cookieOwnerID)
-		} else if cookieOwnerID != session.Get(common.WEBSessionOwnerUinKey) {
-			session.Set(common.WEBSessionOwnerUinKey, cookieOwnerID)
-		}
-		session.Set(common.WEBSessionSupplierID, "0")
-
-		blog.V(5).Infof("skip login, cookie language: %s, cookieOwnerID: %s, rid: %s", webCommon.GetLanguageByHTTPRequest(c), cookieOwnerID, rid)
-		session.Set(common.WEBSessionUinKey, "admin")
-
-		session.Set(common.WEBSessionRoleKey, "1")
-		session.Set(webCommon.IsSkipLogin, "1")
-		if err := session.Save(); err != nil {
-			blog.Warnf("save session failed, err: %s, rid: %s", err.Error(), rid)
-		}
-		return true
-	}
-	user := user.NewUser(config, Engine, CacheCli, LoginPlg)
+	user := user.NewUser(config, Engine, CacheCli)
 	session := sessions.Default(c)
 
 	// check bk_token
@@ -153,12 +129,6 @@ func isAuthed(c *gin.Context, config options.Config) bool {
 	}
 
 	bkTokenName := common.HTTPCookieBKToken
-	if nil != LoginPlg {
-		bkPluginTokenName, err := LoginPlg.Lookup("BKTokenName")
-		if nil == err {
-			bkTokenName = *bkPluginTokenName.(*string)
-		}
-	}
 	bkToken, err := c.Cookie(bkTokenName)
 	blog.V(5).Infof("valid user login session token %s, cookie token %s, rid: %s", ccToken, bkToken, rid)
 	if nil != err || bkToken != ccToken {
