+++ src/web_server/middleware/user/plugins/mananger.go
@@ -26,17 +26,17 @@ func CurrentPlugin(c *gin.Context, version string) metadata.LoginUserPluginInerf
 		version = common.BKDefaultLoginUserPluginVersion
 	}
 
-	var selfPlugin *metadata.LoginPluginInfo
+	var defaultPlugin *metadata.LoginPluginInfo
 	for _, plugin := range manager.LoginPluginInfo {
 		if plugin.Version == version {
 			return plugin.HandleFunc
 		}
 		if common.BKDefaultLoginUserPluginVersion == plugin.Version {
-			selfPlugin = plugin
+			defaultPlugin = plugin
 		}
 	}
-	if nil != selfPlugin {
-		return selfPlugin.HandleFunc
+	if nil != defaultPlugin {
+		return defaultPlugin.HandleFunc
 	}
 
 	return nil
