+++ src/web_server/middleware/user/plugins/method/blueking/userinfo.go
@@ -10,7 +10,7 @@
  * limitations under the License.
  */
 
-package self
+package blueking
 
 import (
 	"encoding/json"
