+++ src/web_server/middleware/user/plugins/register/blueking.go
@@ -13,5 +13,5 @@
 package manager
 
 import (
-	_ "configcenter/src/web_server/middleware/user/plugins/method/self"
+	_ "configcenter/src/web_server/middleware/user/plugins/method/blueking"
 )
